Forandringer i monterte våpen, infanteri og stridsvogner
- La til <LineRenderer> for infanteri og monterte våpen, disse linjene
synes også i builds.
- La til Ammunition.lineWidth
- <MountedWeapon> erstatter nå <TanksShooter> da de gjorde omtrent det
samme
