+++ Assets/Prefabs/Ammunition/7.62 Soviet.prefab
@@ -46,6 +46,7 @@ MonoBehaviour:
   damage: 76
   armorPenetration: 2
   explosionRadius: 0
+  lineWidth: 0.07
   projectilePrefab: {fileID: 0}
 --- !u!1001 &100100000
 Prefab:
