+++ Assets/Prefabs/Weapons/Mk44 Bushmaster II empty.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 136e8dab3cfe0654dba991eb79d434a7
-timeCreated: 1467748137
+guid: b1d7becf0973e874a9db6761c7508674
+timeCreated: 1467634159
 licenseType: Free
 NativeFormatImporter:
   userData:
