+++ Assets/Prefabs/Weapons/Mk44 Bushmaster II.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b1d7becf0973e874a9db6761c7508674
-timeCreated: 1467634159
+guid: 77adec4f68a9b8242a6641d8469efbf0
+timeCreated: 1467747995
 licenseType: Free
 NativeFormatImporter:
   userData:
