+++ Assets/Prefabs/Weapons/Rheinmetall L44 empty.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 136e8dab3cfe0654dba991eb79d434a7
-timeCreated: 1467748137
+guid: 7c4f3503897937b49bbc1a21bcf711ab
+timeCreated: 1467636727
 licenseType: Free
 NativeFormatImporter:
   userData:
