+++ Assets/Prefabs/Weapons/Rheinmetall L44.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7c4f3503897937b49bbc1a21bcf711ab
-timeCreated: 1467636727
+guid: fef28152ee182ab4c88e2da06e008123
+timeCreated: 1467746295
 licenseType: Free
 NativeFormatImporter:
   userData:
