+++ Assets/Scripts/ArtileriShoter.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
 guid: c2ff4ab38d3e3d844a6779381581bf47
-timeCreated: 1464535267
+timeCreated: 1467745320
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: 300
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
