+++ Assets/Scripts/InfantryShooting.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
 guid: 8d28d3104dfce154c9dc52ff1bfbaaf2
-timeCreated: 1464202516
+timeCreated: 1467745320
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: 100
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
