+++ Assets/Scripts/MountedWeapon.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: b7ad7211cf705814db44f42ef402a510
-timeCreated: 1467647766
+guid: 3afcaf85f5cd0ab4ca08ecee18c2ddb3
+timeCreated: 1467745330
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: 400
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
