--- Assets/Scripts/MountedWeaponShooter.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 3afcaf85f5cd0ab4ca08ecee18c2ddb3
-timeCreated: 1467646886
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
