+++ Assets/Scripts/TanksShoter.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
 guid: 714cc38f4ece90548b786faaf4ac3e56
-timeCreated: 1464365749
+timeCreated: 1467745320
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: 200
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
