+++ Assets/Scripts/Unit.cs
@@ -24,6 +24,7 @@ public class Unit : MonoBehaviour
     public int movementSpeed = 5;
     public int rotationSpeed = 120;
     public Weapon weapon;
+    public MountedWeapon[] mountedWeapons;
     public GameObject weaponObject;
     public Sprite icon;
     [HideInInspector]
@@ -40,19 +41,23 @@ public class Unit : MonoBehaviour
 
     void Awake()
     {
-        tag = team.ToString();
+        //unit layeret
+        layermask = 1 << unitLayer;
+
+        //Gir oss noe liv
         currentHealth = startHealth;
 
-        layermask = 1 << unitLayer;
-        if(team == GameManager.team)
-        {
+        //Legger oss inn i listen for riktig lag
+        tag = team.ToString();
+        if (team == GameManager.team)
             GameManager.friendlyUnits.Add(this);
-        }
         else
-        {
             GameManager.enemyUnits.Add(this);
-        }
+
+        //Ser om vi har noen (ekstra) våpen
+        mountedWeapons = GetComponents<MountedWeapon>();
     }
+
     void Start()
     {
         cSelectable = GetComponent<CSelectable>(); cSelectable.unit = this;
@@ -63,32 +68,59 @@ public class Unit : MonoBehaviour
 
         SetWeapon(weapon);
     }
+
     IEnumerator UpdateEnemies()
     {
         Unit enemyUnit;
+
+        //Finner den lengste rekkeviden på alle våpen
+        float maxRange = 0;
+        if (weapon != null)
+            maxRange = weapon.range;
+        for (int i = 0; i < mountedWeapons.Length; i++)
+        {
+            if (mountedWeapons[i].weapon.range > maxRange)
+                maxRange = mountedWeapons[i].weapon.range;
+        }
+
+        //Så lenge vi lever
         while (true)
         {
+            //venter 0.5 sek
             yield return new WaitForSeconds(0.5f);
-            if (weapon == null)
+            //Ikke mye å gjøre om man ikke kan angripe
+            if (maxRange == 0)
                 continue;
 
-            collidersInRange = Physics.OverlapSphere(transform.position, weapon.range, layermask);
+            //Glemmer alle fra forige gang vi sjekket
             enemiesInRange.Clear();
-
             closestEnemy = null;
+
+            //Finner alle collidere i en kule med radiusen maxRande rundt uniten
+            collidersInRange = Physics.OverlapSphere(transform.position, maxRange, layermask);
+            
+            //Distansen til uniten
             float dist;
+            //Distansen til den nærmeste uniten
             float minDist = Mathf.Infinity;
 
+            //For hver collider
             for (int i = 0; i < collidersInRange.Length; i++)
             {
+                //Om det er en fientlig unit
                 enemyUnit = collidersInRange[i].GetComponent<Unit>();
-                if (enemyUnit.team != this.team)
+                if (enemyUnit != null && enemyUnit.team != this.team)
                 {
+                    //Hiver den inn i listen
                     enemiesInRange.Add(enemyUnit);
 
+                    //Finner distansen
                     dist = (enemyUnit.transform.position - this.transform.position).magnitude;
+
+                    //Om det er den bærmeste hittil
                     if (dist < minDist)
                     {
+                        //Husker på det
                         minDist = dist;
                         closestEnemy = enemyUnit;
                     }
@@ -99,18 +131,26 @@ public class Unit : MonoBehaviour
 
     public void SetWeapon(Weapon newWeapon)
     {
+        //Om det ikke er noe driter vi i det
         if (newWeapon == null)
             return;
+
+        //Hvis det har en modell,
         if (newWeapon.loadModel)
         { 
+            //Laster inn prefaben
             GameObject newWeaponObject = GameObject.Instantiate(newWeapon.gameObject);
+            //Setter den på riktig sted
             newWeaponObject.transform.position = weaponObject.transform.position;
             newWeaponObject.transform.parent = weaponObject.transform.parent;
             weaponObject = newWeaponObject;
+            //Henter Weapon komponenten
             weapon = newWeaponObject.GetComponent<Weapon>();
         }
+        //Om ikke
         else
         {
+            //Setter våpenet som det nye
             weapon = newWeapon;
         }
     }
